Refactor most of the unit tests into the tests directory
